Add ReplicaSync trait
